This took quite some time (and is not a good solution...)
Had an issue where I calculated the newValue variables incorrectly in variants. I had just subtracted the value from the old value, which worked fine for one of the two (when the bit was already 1 before I touched). If that didn't happen I ended up with a very different bit layout due to subtracting 2^n when the n-th bit was 0.