Day 20 part 2
Didn't take too long, but I slowed down a bit by the difficulty of debugging any issues (I had swapped the indices in the tile access in lineFromTile). That's probably an intended part of the problem, but it still doesn't make for a fun time. I still found it before I got too annoyed, though