Day 21 part 1
Went alright, I got stuck a bit on having counted the occurences of allergens (5), which happens to line up with the number of occurences of ingredients that do not contain allergens.
It wasn't the first place I looked when debugging why the example worked but not the actual input, thinking instead that I likely had some issue in the algorithm.